update
funcao download e manutencao arquivo temporario

+++ R/candidate_fed.R
@@ -21,6 +21,8 @@
 #' when \code{ascii = TRUE}.
 #' 
 #' @param export (\code{logical}). Should the downloaded data be saved in .dta and .sav in the current directory?
+#' 
+#' @param temp (\code{logical}). If \code{TRUE} keep temporary compressed file
 #'
 #' @details If export is set to \code{TRUE}, the downloaded data is saved as .dta and .sav
 #'  files in the current directory.
@@ -112,7 +114,9 @@
 #' df <- candidate_fed(2002)
 #' }
 
-candidate_fed <- function(year, uf = "all", br_archive = FALSE, ascii = FALSE, encoding = "latin1", export = FALSE){
+candidate_fed <- function(year, uf = "all", br_archive = FALSE,
+                          ascii = FALSE, encoding = "latin1", 
+                          export = FALSE, temp = TRUE){
 
 
   # Input tests
@@ -121,14 +125,17 @@ candidate_fed <- function(year, uf = "all", br_archive = FALSE, ascii = FALSE, e
   uf <- test_uf(uf)
   test_br(br_archive)
 
-  # Download the data
-  dados <- tempfile()
-  sprintf("http://agencia.tse.jus.br/estatistica/sead/odsele/consulta_cand/consulta_cand_%s.zip", year) %>%
-    download.file(dados)
-  unzip(dados, exdir = paste0("./", year))
-  unlink(dados)
-
-  message("Processing the data...")
+  filenames  <- paste0("/consulta_cand_", year, ".zip")
+  dados <- paste0(file.path(tempdir()), filenames)
+  url <- "https://cdn.tse.jus.br/estatistica/sead/odsele/consulta_cand%s"
+  
+  # Downloads the data
+  download_unzip(url, dados, filenames, year)
+  
+  # remover temp file
+  if(temp == FALSE){
+    unlink(dados)
+  }
 
   # Cleans the data
   setwd(as.character(year))
