+++ R/candidate_local.R
@@ -17,6 +17,8 @@
 #' when \code{ascii = TRUE}.
 #' 
 #' @param export (\code{logical}). Should the downloaded data be saved in .dta and .sav in the current directory?
+#' 
+#' @param temp (\code{logical}). If \code{TRUE} keep temporary compressed file
 #'
 #' @details If export is set to \code{TRUE}, the downloaded data is saved as .dta and .sav
 #'  files in the current directory.
@@ -86,22 +88,27 @@
 #' df <- candidate_local(2000)
 #' }
 
-candidate_local <- function(year, uf = "all", ascii = FALSE, encoding = "latin1", export = FALSE){
+candidate_local <- function(year, uf = "all", ascii = FALSE,
+                            encoding = "latin1", export = FALSE,
+                            temp = TRUE){
 
   # Input tests
   test_encoding(encoding)
   test_local_year(year)
   uf <- test_uf(uf)
   
+  filenames  <- paste0("/consulta_cand_", year, ".zip")
+  dados <- paste0(file.path(tempdir()), filenames)
+  url <- "https://cdn.tse.jus.br/estatistica/sead/odsele/consulta_cand%s"
+  
   # Downloads the data
-  dados <- tempfile()
-  sprintf("http://agencia.tse.jus.br/estatistica/sead/odsele/consulta_cand/consulta_cand_%s.zip", year) %>%
-    download.file(dados)
-  unzip(dados, exdir = paste0("./", year))
-  unlink(dados)
-
-  message("Processing the data...")
-
+  download_unzip(url, dados, filenames, year)
+  
+  # remover temp file
+  if(temp == FALSE){
+    unlink(dados)
+  }
+  
   # Cleans the data
   setwd(as.character(year))
   banco <- juntaDados(uf, encoding, FALSE)
